honeypot/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
motor>=3.4.0
langchain>=0.1.16
langchain-core>=0.1.42
langchain-community>=0.0.32
langchain-groq>=0.1.2
langgraph>=0.0.35
google-generativeai>=0.5.0
python-dotenv>=1.0.1
httpx>=0.27.0
email-validator>=2.1.1
dnspython>=2.6.1
langchain-google-genai>=1.0.1
uv

# Voice Processing Dependencies
faster-whisper>=1.0.0
piper-tts>=1.2.0
pydub>=0.25.1
numpy>=1.24.0
soundfile>=0.12.1
python-multipart>=0.0.6
pyttsx3>=2.90
gTTS>=2.5.1

# Live Takeover Dependencies
websockets>=12.0
python-socketio>=5.11.0
reportlab>=4.1.0
python-whois>=0.9.4

# JWT Authentication (Fix 3)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# S3-compatible Object Storage (Fix 4)
minio>=7.2.0

# Rate Limiting (Fix 5)
slowapi>=0.1.9
redis>=5.0.0